requirements.txt
aniso8601==9.0.1
anyio==3.7.1
backoff==2.2.1
black==23.7.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.6
click==8.1.6
colorama==0.4.6
coloredlogs==15.0.1
configura==0.1
exceptiongroup==1.1.3
fastapi==0.99.1
filelock==3.12.2
Flask==2.3.2
Flask-RESTful==0.3.10
flatbuffers==23.5.26
fsspec==2023.6.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
loguru==0.7.0
MarkupSafe==2.1.3
marshmallow==3.20.1
monotonic==1.6
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.25.2
onnxruntime==1.15.1
overrides==7.4.0
packaging==23.1
pathspec==0.11.2
peewee==3.16.3
Pillow==10.0.0
platformdirs==3.10.0
posthog==3.0.2
protobuf==4.24.1
pulsar-client==3.2.0
pydantic==1.10.12
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.66.1
transformers==4.31.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.7
win32-setctime==1.1.0
